refactor: 디렉토리 이동, 어노테이션 분리
